ImageMatcher/app-data/requirements.txt
numpy==1.26.4
rasterio==1.3.10
pandas==2.2.2
matplotlib==3.9.0
scikit-image==0.23.2
pillow==10.3.0
opencv-python-headless==4.10.0.82
uvicorn==0.30.1
fastapi==0.111.0